IGListAdapter: comment out IGAssertion causing crash
Debug builds of IGListAdapter enable NSFoundation assertions in contrast with Adhoc and Release builds which do not. Crashes can occur using a Debug build when a nil IGListSectionController is encountered.

A suspected cause for the nil IGListSectionController crash may lie in a race condition between the UICollectionViewDelegate method [UICollectionViewDelegate collectionView:didEndDisplayingCell:forItemAtIndexPath:] and the UICollectionView property [UICollectionView visibleCells]. IGListSectionControllers are removed when the didEndDisplayCell delegate method is called. However, [UICollectionView visibleCells] may in an indeterministic fashion return UICollectionCells correlating to removed IGListSectionControllers. This inconsistency is responsible for a crash as [IGListAdapter visibleCellsForSectionController:] does not expects [UICollectionView visibleCells] to report UICollectionViewCells that have reported the didEndDisplayCell delegate call.